Revert "https://myddleware.atlassian.net/browse/PROJ-647 transfer all the formula code to formula-wizard.js"
This reverts commit e07e41d71fda23a86f98a657dbf27299f7946bcf.
