+++ assets/js/regle.js
@@ -1729,6 +1729,71 @@ function openFormula() {
   });
 }
 
+function colorationSyntax() {
+  $("#area_color").html($("#area_insert").val());
+
+  if ($("#area_insert").val() == "") {
+    $("#area_color").empty();
+  }
+
+  $("#area_color").each(function () {
+    var text = $(this).html();
+
+    //---
+    text = text.replace(/\===/g, "[begin] class='operateur'[f]===[end]");
+    text = text.replace(/\==/g, "[begin] class='operateur'[f]==[end]");
+    text = text.replace(/\!==/g, "[begin] class='operateur'[f]!==[end]");
+    text = text.replace(/\!=/g, "[begin] class='operateur'[f]!=[end]");
+    //---
+    text = text.replace(/\!/g, "[begin] class='operateur'[f]![end]");
+    text = text.replace(/\./g, "[begin] class='operateur'[f].[end]");
+    text = text.replace(/\:/g, "[begin] class='operateur'[f]:[end]");
+    text = text.replace(/\?/g, "[begin] class='operateur'[f]?[end]");
+    text = text.replace(/\(/g, "[begin] class='operateur'[f]([end]");
+    text = text.replace(/\)/g, "[begin] class='operateur'[f])[end]");
+    //--
+    text = text.replace(/\//g, "[begin] class='operateur'[f]/[end]");
+    text = text.replace(/\+/g, "[begin] class='operateur'[f]+[end]");
+    text = text.replace(/\-/g, "[begin] class='operateur'[f]-[end]");
+    text = text.replace(/\*/g, "[begin] class='operateur'[f]*[end]");
+    //---
+    text = text.replace(/\>=/g, "[begin] class='operateur'[f]>=[end]");
+    text = text.replace(/\>/g, "[begin] class='operateur'[f]>[end]");
+    text = text.replace(/\<=/g, "[begin] class='operateur'[f]<=[end]");
+    text = text.replace(/\</g, "[begin] class='operateur'[f]<[end]");
+    //---
+    text = text.replace(/\{/g, "[begin] class='variable'[f]{");
+    text = text.replace(/\}/g, "}[end]");
+    //---
+    text = text.replace(
+      /\"([\s\S]*?)\"/g,
+      "[begin] class='chaine'[f]\"$1\"[end]"
+    );
+    //---
+    text = text.replace(/\[begin\]/g, "<span");
+    text = text.replace(/\[f\]/g, ">");
+    text = text.replace(/\[end\]/g, "</span>");
+    $("#area_color").html(text);
+  });
+
+  // supprime les doublons
+  $(".operateur", "#area_color").each(function () {
+    if ($(this).parent().attr("class") == "chaine") {
+      $(this).before($(this).html());
+      $(this).remove();
+    }
+  });
+
+  // supprime les doublons
+  $(".variable", "#area_color").each(function () {
+    if ($(this).parent().attr("class") == "chaine") {
+      $(this).before($(this).html());
+      $(this).remove();
+    }
+  });
+}
+//-- syntax color
+
 // ---- AJOUT CHAMP CIBLE  ---------------------------------------------------------------
 
 function existField(name) {
@@ -2231,4 +2296,274 @@ document.addEventListener('DOMContentLoaded', function () {
   tooltipTriggerList.forEach(tooltipTriggerEl => {
       new bootstrap.Tooltip(tooltipTriggerEl);
   });
+});
+
+// Function wizard handling
+$(document).ready(function() {
+    const functionSelect = $('#function-select');
+    const lookupOptions = $('#lookup-options');
+    const lookupRule = $('#lookup-rule');
+    const lookupField = $('#lookup-field');
+    const flagFunctionWizardEnd = $('#flag-function-wizard-end');
+    const functionParameter = $('#function-parameter');
+    const insertFunctionBtn = $('#insert-function-parameter');
+    let tooltipVisible = false; // Changed to false by default
+    let currentTooltip = '';
+    let selectedFunction = '';
+    
+    // Handle tooltip toggle button
+    $('#toggle-tooltip').on('click', function() {
+        tooltipVisible = !tooltipVisible;
+        const tooltipBox = $('#function-tooltip');
+        
+        if (tooltipVisible) {
+            $(this).find('i').removeClass('fa-question').addClass('fa-question-circle');
+            if (functionSelect.val() && currentTooltip) {
+                tooltipBox.text(currentTooltip).show();
+            }
+        } else {
+            $(this).find('i').removeClass('fa-question-circle').addClass('fa-question');
+            tooltipBox.hide();
+        }
+    });
+
+    // Show tooltip when option changes
+    functionSelect.on('change', function() {
+        const selectedOption = $(this).find('option:selected');
+        const tooltip = selectedOption.data('tooltip');
+        const tooltipBox = $('#function-tooltip');
+        
+        selectedFunction = $(this).val();
+        currentTooltip = tooltip;
+
+        // if the selected function is a mwd function, we need to hide the function-parameter input, else we need to show it
+        if (selectedFunction.startsWith('mdw_')) {
+            $('#function-parameter').hide();
+        } else {
+            $('#function-parameter').show();
+        }
+        
+        // Only show tooltip if tooltipVisible is true and a function is selected
+        if (tooltip && tooltipVisible && $(this).val()) {
+            tooltipBox.text(tooltip).show();
+        } else {
+            tooltipBox.hide();
+        }
+
+        if (selectedFunction === 'round') {
+          $('#round-precision-input').show();
+      } else {
+          $('#round-precision-input').hide();
+      }
+
+        if (selectedFunction === 'lookup') {
+            lookupOptions.show();
+            $('#function-parameter-input').hide(); // Hide parameter input for lookup
+            
+            // Populate rules dropdown
+            $.ajax({
+                url: lookupgetrule,
+                method: 'GET',
+                data: {
+                    arg1: connectorsourceidlookup,
+                    arg2: connectortargetidlookup
+                },
+                success: function(rules) {
+                    lookupRule.empty();
+                    lookupRule.append('<option value="">' + translations.selectRule + '</option>');
+                    rules.forEach(rule => {
+                        lookupRule.append(`<option value="${rule.id}">${rule.name}</option>`);
+                    });
+                    lookupRule.prop('disabled', false);
+                }
+            });
+        } else {
+            lookupOptions.hide();
+            $('#function-parameter-input').show(); // Show parameter input for other functions
+        }
+    });
+
+    // Handle parameter insertion
+    insertFunctionBtn.on('click', function() {
+        if (!selectedFunction) return; // Do nothing if no function is selected
+        
+        // Get the function category from the selected option
+        const functionCategory = $('#function-select option:selected').data('type');
+
+        if (selectedFunction === 'round') {
+          const parameterValue = functionParameter.val().trim();
+          const precisionInput = $('#round-precision');
+          const precision = parseInt(precisionInput.val());
+          
+          // Validate precision
+          if (isNaN(precision) || precision < 1 || precision > 100) {
+              precisionInput.addClass('is-invalid');
+              return;
+          }
+          
+          precisionInput.removeClass('is-invalid');
+          
+          const areaInsert = $('#area_insert');
+          const position = areaInsert.getCursorPosition();
+          const content = areaInsert.val();
+          
+          // Construct round function call with precision
+          const functionCall = `round(${parameterValue}, ${precision})`;
+          
+          const newContent = 
+              content.substr(0, position) +
+              functionCall +
+              content.substr(position);
+              
+          areaInsert.val(newContent);
+          
+          // Clear inputs
+          functionParameter.val('');
+          precisionInput.val('');
+        
+        // Special handling for MDW functions
+        } else if (selectedFunction.startsWith('mdw_')) {
+            const areaInsert = $('#area_insert');
+            const position = areaInsert.getCursorPosition();
+            const content = areaInsert.val();
+
+
+            
+            // For MDW functions, just insert the function name as a string
+            const functionCall = `"${selectedFunction}"`;
+            
+            const newContent = 
+                content.substr(0, position) +
+                functionCall +
+                content.substr(position);
+                
+            areaInsert.val(newContent);
+        } else {
+            // Normal function handling
+            const parameterValue = functionParameter.val().trim();
+            const areaInsert = $('#area_insert');
+            const position = areaInsert.getCursorPosition();
+            const content = areaInsert.val();
+            
+            // Create the function call based on category
+            let functionCall = '';
+            if (parameterValue) {
+                switch(functionCategory) {
+                    case 1: // mathematical
+                        functionCall = `${selectedFunction}(${parameterValue})`; // No quotes for numbers
+                        break;
+                    case 2: // text
+                    case 3: // date
+                        functionCall = `${selectedFunction}("${parameterValue}")`; // Add quotes for text and dates
+                        break;
+                    case 4: // constant
+                        functionCall = `${selectedFunction}()`; // Constants don't need parameters
+                        break;
+                    default:
+                        functionCall = `${selectedFunction}("${parameterValue}")`; // Default to quoted parameter
+                }
+            } else {
+                functionCall = `${selectedFunction}()`;
+            }
+            
+            const newContent = 
+                content.substr(0, position) +
+                functionCall +
+                content.substr(position);
+                
+            areaInsert.val(newContent);
+        }
+        
+        // Clear the parameter input
+        functionParameter.val('');
+        
+        // Update syntax highlighting
+        colorationSyntax();
+        theme(style_template);
+    });
+
+    // When a rule is selected
+    lookupRule.on('change', function() {
+        const selectedRule = $(this).val();
+        
+        if (selectedRule) {
+            lookupField.empty();
+            lookupField.append('<option value="">' + translations.selectField + '</option>');
+            
+            // Get fields from the existing select element
+            $('#champs_insert option').each(function() {
+                const fieldName = $(this).val();
+                lookupField.append(`<option value="${fieldName}">${fieldName}</option>`);
+            });
+            
+            lookupField.prop('disabled', false);
+        } else {
+            lookupField.prop('disabled', true);
+        }
+    });
+
+    // Remove the lookupField.on('change') handler since we don't want automatic insertion
+    lookupField.off('change');
+
+    // Add handler for submit lookup button
+    $('#submit-lookup').on('click', function() {
+        const selectedField = lookupField.find('option:selected');
+        if (selectedField.val()) {
+            const fieldName = selectedField.text().split(' (')[0];
+            let errorEmpty = $('#lookup-error-empty').is(':checked');
+            let errorNotFound = $('#lookup-error-not-found').is(':checked');
+
+            // Convert boolean to 1/0
+            errorEmpty = errorEmpty ? 1 : 0;
+            errorNotFound = errorNotFound ? 1 : 0;
+            
+            // Construct the complete lookup formula
+            const lookupFormula = `lookup({${fieldName}}, "${lookupRule.val()}", ${errorEmpty}, ${errorNotFound})`;
+            
+            const areaInsert = $('#area_insert');
+            const position = areaInsert.getCursorPosition();
+            const content = areaInsert.val();
+            
+            const newContent = 
+                content.substr(0, position) +
+                lookupFormula +
+                content.substr(position);
+                
+            areaInsert.val(newContent);
+            colorationSyntax();
+            theme(style_template);
+        }
+    });
+
+    $('#round-precision').on('input', function() {
+      const value = this.value;
+      
+      // Remove any non-digit characters
+      const sanitizedValue = value.replace(/[^0-9]/g, '');
+      
+      if (sanitizedValue !== value) {
+          this.value = sanitizedValue;
+      }
+      
+      const precision = parseInt(sanitizedValue);
+      
+      if (isNaN(precision) || precision < 1 || precision > 100) {
+          $(this).addClass('is-invalid');
+      } else {
+          $(this).removeClass('is-invalid');
+      }
+  });
+
+// wait for the press of the close button which is #area_quit
+$('#area_quit').on('click', function() {
+    resetFunctionWizard();
+});
+
+// function to reset the function wizard
+function resetFunctionWizard() {
+    $('#function-select').val('').trigger('change');
+    $('#lookup-rule').val('').trigger('change');
+    $('#lookup-field').val('').prop('disabled', true);
+}
+
 });
